One can change user avatar

+++ SocialPhotoEditor.DataLayer/Repositories/EditedRepositories/ChangedRepositories/Implementations/CommentRepository.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using System.Security.Cryptography.X509Certificates;
 using SocialPhotoEditor.DataLayer.DatabaseContextes;
 using SocialPhotoEditor.DataLayer.DatabaseModels;
 
@@ -29,6 +28,9 @@ namespace SocialPhotoEditor.DataLayer.Repositories.EditedRepositories.ChangedRep
         {
             using (var db = new ApplicationDbContext())
             {
+                data = db.Comments.FirstOrDefault(x => x.CommentatorId == data.CommentatorId && x.ImageId == data.ImageId && x.Time == data.Time);
+                if (data == null)
+                    return;
                 db.Comments.Remove(data);
                 db.SaveChanges();
             }
